requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==20.3.0       
certifi==2020.12.5  
chardet==4.0.0      
cogwatch==2.1.0     
discord.py==1.7.1   
flake8==3.9.1       
idna==2.10
Jinja2==2.11.3
MarkupSafe==1.1.1
mccabe==0.6.1
multidict==5.1.0
numpy==1.20.2
pandas==1.2.4
pycodestyle==2.7.0
pyflakes==2.3.1
pygicord==0.4.1
python-dateutil==2.8.1
python-dotenv==0.17.0
pytz==2021.1
requests==2.25.1
six==1.15.0
tabulate==0.8.9
typing-extensions==3.7.4.3
unicode==2.8
Unidecode==1.2.0
urllib3==1.26.4
watchgod==0.7
yarl==1.6.3